Documentation updated. (Falta descompactar LZ)
