+++ medusa/local_activation/nonlinear_parameters.py
@@ -12,273 +12,6 @@ from scipy.signal import decimate
 from medusa.components import ThreadWithReturnValue
 from medusa.utils import check_dimensions
 
-
-def central_tendency_measure(signal, r):
-    """
-    This method implements the central tendency measure (CTM). This parameter is
-    useful to quantify the variability of a signal. It is based on calculating
-    the second-order differences diagram of the time series and then counting
-    the points within a radius "r". CTM assigns higher values to less variable
-    signals
-
-    References
-    ----------
-    Cohen, M. E., Hudson, D. L., & Deedwania, P. C. (1996). Applying
-    continuous chaotic modeling to cardiac signal analysis. IEEE Engineering in
-    Medicine and Biology Magazine, 15(5), 97-102.
-
-    Parameters
-    ----------
-    signal : numpy.ndarray
-        Signal with shape [n_epochs, n_samples, n_channels].
-    r : double
-        Radius used to compute the CTM.
-
-    Returns
-    -------
-    ctm : numpy.ndarray
-        CTM values for each channel in "signal". [n_epochs, n_channels].
-    """
-
-    #  Error check
-    if not np.issubdtype(signal.dtype, np.number):
-        raise ValueError('data matrix contains non-numeric values')
-
-    # Check dimensions
-    signal = check_dimensions(signal)
-
-    # Signal dimensions
-    n_epo = signal.shape[0]
-    n_samp = signal.shape[1]
-    n_cha = signal.shape[2]
-
-    # Values within a range (mean +- 3 std)
-    upper_bound = np.mean(signal, axis=1) + 3 * np.std(signal, axis=1)
-    lower_bound = np.mean(signal, axis=1) - 3 * np.std(signal, axis=1)
-    idx_within_range = np.logical_and((signal < upper_bound[:, None, :]),
-                                      (signal > lower_bound[:, None, :]))
-    idx_out_upper = (signal > upper_bound[:, None, :])
-    idx_out_lower = (signal < lower_bound[:, None, :])
-
-    # Maximum value in the above defined range
-    max_value = np.empty((n_epo, n_cha))
-    for ep in range(n_epo):
-        for ch in range(n_cha):
-            max_value[ep, ch] = np.max(
-                abs(signal[ep, idx_within_range[ep, :, ch], ch]), axis=0)
-
-    # Normalize the values within the range by its maximum.Values above that
-    # range will be 1, and below the range will be - 1
-    data_norm = np.zeros_like(signal)
-    data_norm[idx_within_range] = np.divide(
-        signal[idx_within_range],
-        np.tile(max_value, (1, n_samp, 1)).flatten()[
-            idx_within_range.flatten()])
-    data_norm[idx_out_upper] = 1
-    data_norm[idx_out_lower] = -1
-
-    # Difference time series
-    y = data_norm[:, 3:n_samp, :] - data_norm[:, 2:n_samp - 1, :]
-    x = data_norm[:, 2:n_samp - 1, :] - data_norm[:, 1:n_samp - 2, :]
-
-    # CTM - Values below the radius 'r'
-    ctm = np.sum(np.sqrt(np.square(x) + np.square(y)) < r, axis=1) / (
-                n_samp - 2)
-
-    return ctm
-
-
-def sample_entropy(signal, m, r, dist_type='chebyshev'):
-    """ This method implements the sample entropy (SampEn). SampEn is an
-    irregularity measure that assigns higher values to more irregular time
-    sequences. It has two tuning parameters: the sequence length (m) and the
-    tolerance (r)
-
-    Notes: IF A = 0 or B = 0, SamEn would return an infinite value.
-    However, the lowest non-zero conditional probability that SampEn should
-    report is A/B = 2/[(N-m-1)*(N-m)].
-    SampEn has the following limits:
-                 - Lower bound: 0
-                 - Upper bound : log(N-m) + log(N-m-1) - log(2)
-
-    References
-    ----------
-    Richman, J. S., & Moorman, J. R. (2000). Physiological
-    time-series analysis using approximate entropy and sample entropy. American
-    Journal of Physiology-Heart and Circulatory Physiology.
-
-    Parameters
-    ----------
-    signal : numpy.ndarray
-        Signal with shape [n_epochs, n_samples, n_channels].
-    m : int
-        Sequence length
-    r : float
-        Tolerance
-    dist_type : string
-        Distance metric
-
-    Returns
-    -------
-    sampen : numpy.ndarray
-        SampEn value. [n_epochs, n_channels]
-    """
-    # Check dimensions
-    signal = check_dimensions(signal)
-
-    # Check Errors
-    if m > signal.shape[1]:
-        raise ValueError('Embedding dimension must be smaller than the signal '
-                         'length (m<N).')
-    if not isinstance(dist_type, str):
-        raise ValueError('Distance type must be a string.')
-    if dist_type not in ['braycurtis', 'canberra', 'chebyshev', 'cityblock',
-                         'correlation', 'cosine', 'dice', 'euclidean',
-                         'hamming', 'jaccard', 'jensenshannon', 'kulsinski',
-                         'mahalanobis', 'matching', 'minkowski',
-                         'rogerstanimoto', 'russellrao', 'seuclidean',
-                         'sokalmichener', 'sokalsneath', 'sqeuclidean', 'yule']:
-        raise ValueError(
-            'Distance type unknown. Please, check allowed distances'
-            'in pdist function from scipy.spatial.distance module.')
-
-    # Useful parameters
-    n_epo = signal.shape[0]
-    N = signal.shape[1]
-    n_channels = signal.shape[2]
-    sigma = np.std(signal, axis=1)
-    templates_m = []
-    templates_m_plus_one = []
-    B, A, value = np.empty((n_epo, n_channels)), np.empty((n_epo, n_channels)),\
-                  np.empty((n_epo, n_channels))
-
-    # Calculate B values
-    for i in range(N - m + 1):
-        templates_m.append(signal[:, i:i + m, :])
-    templates_m = np.array(templates_m)
-    for e_idx in range(n_epo):
-        w_threads = []
-        for ch_idx in range(n_channels):
-            t = ThreadWithReturnValue(
-                target=pdist,
-                args=(templates_m[:, e_idx, :, ch_idx], dist_type,))
-            w_threads.append(t)
-            t.start()
-        for th_idx, thread in enumerate(w_threads):
-            B[e_idx, th_idx] = np.sum(thread.join() <= sigma[0, th_idx] * r)
-
-    # Check if there is any B = 0
-    zeros_idx = np.where(B == 0)
-    value[zeros_idx] = math.inf
-
-    # Calculate A values
-    m += 1
-    for i in range(N - m + 1):
-        templates_m_plus_one.append(signal[:, i:i + m, :])
-    templates_m_plus_one = np.array(templates_m_plus_one)
-    for e_idx in range(n_epo):
-        w_threads = []
-        for ch_idx in range(n_channels):
-            t = ThreadWithReturnValue(
-                target=pdist, args=(templates_m_plus_one[:, e_idx, :, ch_idx],
-                                    dist_type,))
-            w_threads.append(t)
-            t.start()
-        for th_idx, thread in enumerate(w_threads):
-            A[e_idx, th_idx] = np.sum(thread.join() <= sigma[0, th_idx] * r)
-
-    # Check if there is any A = 0
-    zeros_idx = np.where(A == 0)
-    value[zeros_idx] = math.inf
-
-    non_inf = np.where(value != math.inf)
-    value[non_inf] = -np.log((A[non_inf] / B[non_inf])*((N - m + 1) / (N - m - 1)))
-
-    # If there is infinity values
-    inf_indx = np.where(value == math.inf)
-    value[inf_indx] = -np.log(2 / ((N - m - 1) * (N - m)))
-    return value
-
-
-def multiscale_entropy(signal, max_scale, m, r):
-    """ This method implements the Multiscale Entropy (MSE). MSE is a
-    measurement of complexity which measures the irregularity of a signal over
-    multiple time-scales. This is accomplished through estimation of the Sample
-    Entropy (SampEn) on coarse-grained versions of the original signal. As a
-    result of the these calculations, MSE curves are obtained and can be used to
-    compare the complexity of time-series. The MSE curve whose entropy values
-    are higher for the most of time-scales is considered more complex.
-
-    References
-    ----------
-    Costa, M., Goldberger, A. L., & Peng, C. K. (2005). Multiscale entropy
-    analysis of biological signals, Physical review E, 71(2), 021906.
-
-    Parameters
-    ----------
-    signal : numpy.ndarray
-        MEEG Signal. [n_epochs, n_samples, n_channels].
-    max_scale : int
-        Maximum scale value
-    m : int
-        Sequence length
-    r : float
-        Tolerance
-
-    Returns
-    -------
-    mse_result : numpy.ndarray
-        SampEn values for each scale for each channel in "signal" .
-        [n_epochs, max_scale, n_channels].
-
-    """
-    # Check dimensions
-    signal = check_dimensions(signal)
-
-    # Signal dimensions
-    n_epo = signal.shape[0]
-    n_channels = signal.shape[2]
-
-    mse_result = np.empty((n_epo, max_scale, n_channels))
-    w_threads = list()
-    scales = list()
-    for i in range(1, max_scale + 1):
-        if i == 1:
-            t = ThreadWithReturnValue(
-                target=sample_entropy,
-                args=(signal, m, r, 'chebyshev'))
-        else:
-            t = ThreadWithReturnValue(
-                target=sample_entropy,
-                args=(__coarse_grain(signal, i),
-                      m, r, 'chebyshev'))
-        w_threads.append(t)
-        scales.append(i)
-        t.start()
-    for t_idx, thread in enumerate(w_threads):
-        mse_result[:, t_idx, :] = thread.join()
-    return mse_result
-
-
-def __coarse_grain(signal, scale, decimate_mode=True):
-    if decimate_mode:
-        return decimate(signal, scale, axis=1)
-    else:
-        # Signal dimensions
-        n_epo = signal.shape[0]
-        N = signal.shape[1]
-        n_cha = signal.shape[2]
-
-        # Number of coarse grains in which the signal is split
-        tau = int(round(N / scale))
-
-        # Returned signal
-        y = np.empty((n_epo,tau,n_cha))
-        for i in range(tau):
-            y[:, i, :] = np.mean(signal[:, i * scale:(i * scale + scale), :])
-        return y
-
-
 def __lempelziv_complexity(signal):
     """ Calculate the  signal binarisation and the Lempel-Ziv's complexity. This
     version allows the algorithm to be used for signals with more than one
@@ -368,7 +101,6 @@ def lempelziv_complexity(signal):
                  int(n_sample))
         lz_result[e_idx, :] = lz_channel_values
     return lz_result
-
 
 def multiscale_lempelziv_complexity(signal, W):
     """Calculate the multiscale signal binarisation and the Multiscale
